gh-151722: Move frozendict free-threading regression test into test_dict.py
